Put load of fixtures in dedicated class for integration tests
