+++ src/Pim/Bundle/ResearchBundle/tests/fixtures/EntityLoader/EntityFixtureLoader.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pim\Bundle\ResearchBundle\tests\fixtures\EntityLoader;
+
+interface EntityFixtureLoader
+{
+    public function load($entity): void;
+}
