+++ src/Pim/Bundle/ResearchBundle/tests/integration/Infrastructure/Persistence/Database/DatabaseAttributeRepositoryTestCase.php
@@ -6,6 +6,7 @@ use Doctrine\DBAL\Types\Type;
 use PHPUnit\Framework\Assert;
 use Pim\Bundle\ResearchBundle\DomainModel\Attribute\Attribute;
 use Pim\Bundle\ResearchBundle\DomainModel\Attribute\AttributeCode;
+use Pim\Bundle\ResearchBundle\tests\fixtures\EntityLoader\Database\AttributeLoader;
 use Pim\Bundle\ResearchBundle\tests\fixtures\ResetDatabase;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
@@ -17,7 +18,8 @@ class DatabaseAttributeRepositoryTestCase extends KernelTestCase
     protected function setUp()
     {
         static::bootKernel(['debug' => false]);
-        (new ResetDatabase(static::$kernel->getContainer()->get('doctrine.orm.entity_manager')))();
+        $entityManager = static::$kernel->getContainer()->get('doctrine.orm.entity_manager');
+        (new ResetDatabase($entityManager))();
 
         $attribute1 = new Attribute(
             AttributeCode::createFromString('attribute_code_1'),
@@ -33,8 +35,9 @@ class DatabaseAttributeRepositoryTestCase extends KernelTestCase
             false
         );
 
-        $this->persistAttributeInDatabase($attribute1);
-        $this->persistAttributeInDatabase($attribute2);
+        $attributeLoader = new AttributeLoader($entityManager);
+        $attributeLoader->load($attribute1);
+        $attributeLoader->load($attribute2);
     }
 
     public function test_with_code_on_persisted_attribute()
@@ -88,53 +91,5 @@ class DatabaseAttributeRepositoryTestCase extends KernelTestCase
             AttributeCode::createFromString('foo'),
         ]);
         Assert::assertEquals([], $attributes);
-    }
-
-    private function persistAttributeInDatabase(Attribute $attribute): void
-    {
-        $entityManager = static::$kernel->getContainer()->get('doctrine.orm.entity_manager');
-
-        $sql = <<<SQL
-            INSERT INTO pim_catalog_attribute (
-                code, 
-                attribute_type, 
-                is_localizable,
-                is_scopable, 
-                sort_order, 
-                is_required,
-                is_unique,
-                entity_type,
-                backend_type,
-                created,
-                updated
-            )
-            VALUES (
-                :code,
-                :attribute_type,
-                :is_localizable,
-                :is_scopable,
-                :sort_order,
-                :is_required,
-                :is_unique,
-                :entity_type,
-                :backend_type,
-                :created,
-                :updated
-            )
-SQL;
-
-        $stmt = $entityManager->getConnection()->prepare($sql);
-        $stmt->bindValue('code', $attribute->code()->getValue(), Type::STRING);
-        $stmt->bindValue('attribute_type', $attribute->type(), Type::STRING);
-        $stmt->bindValue('is_localizable', $attribute->localizable(), Type::BOOLEAN);
-        $stmt->bindValue('is_scopable', $attribute->scopable(), Type::BOOLEAN);
-        $stmt->bindValue('sort_order', 1, Type::INTEGER);
-        $stmt->bindValue('is_required', true, Type::BOOLEAN);
-        $stmt->bindValue('is_unique', true, Type::BOOLEAN);
-        $stmt->bindValue('entity_type', 'Pim\Component\Catalog\Model\Product\'', Type::STRING);
-        $stmt->bindValue('backend_type', 'text', Type::STRING);
-        $stmt->bindValue('created', new \DateTime(), Type::DATETIME);
-        $stmt->bindValue('updated', new \DateTime(), Type::DATETIME);
-        $stmt->execute();
     }
 }
